+++ Harden-Windows-Security Module/Main files/C#/GUI/Main/FindVisualChild.cs
@@ -1,0 +1,35 @@
+using System;
+using System.Windows;
+using System.Windows.Media;
+
+#nullable enable
+
+namespace HardenWindowsSecurity
+{
+    public partial class GUIMain
+    {
+        /// <summary>
+        /// Find a visual child of a parent object
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="parent"></param>
+        /// <returns></returns>
+        public T? FindVisualChild<T>(DependencyObject parent) where T : DependencyObject
+        {
+            for (int i = 0; i < VisualTreeHelper.GetChildrenCount(parent); i++)
+            {
+                var child = VisualTreeHelper.GetChild(parent, i);
+                if (child is T tChild)
+                {
+                    return tChild;
+                }
+                var result = FindVisualChild<T>(child);
+                if (result != null)
+                {
+                    return result;
+                }
+            }
+            return null;
+        }
+    }
+}
