+++ Harden-Windows-Security Module/Main files/C#/Others/ActivityTracker.cs
