+++ Harden-Windows-Security Module/Main files/C#/Others/AsyncDownloader.cs
