+++ Harden-Windows-Security Module/Main files/C#/Others/CategoryProcessing.cs
