+++ Harden-Windows-Security Module/Main files/C#/Others/ConditionalResultAdd.cs
