+++ Harden-Windows-Security Module/Main files/C#/Others/ControlledFolderAccessHandler.cs
