+++ Harden-Windows-Security Module/Main files/C#/Others/CultureInfoHelper.cs
