+++ Harden-Windows-Security Module/Main files/C#/Others/EccCurveComparer.cs
