+++ Harden-Windows-Security Module/Main files/C#/Others/ExportSecurityPolicy.cs
