+++ Harden-Windows-Security Module/Main files/C#/Others/ForceRelocateImagesForFiles.cs
