+++ Harden-Windows-Security Module/Main files/C#/Others/GetLocalUser.cs
