+++ Harden-Windows-Security Module/Main files/C#/Others/GetMDMResultValue.cs
