+++ Harden-Windows-Security Module/Main files/C#/Others/GetOneDriveDirectories.cs
