+++ Harden-Windows-Security Module/Main files/C#/Others/GitExesFinder.cs
