+++ Harden-Windows-Security Module/Main files/C#/Others/GitHubDesktopFinder.cs
