+++ Harden-Windows-Security Module/Main files/C#/Others/GlobalVars.cs
