+++ Harden-Windows-Security Module/Main files/C#/Others/HashtableChecker.cs
