+++ Harden-Windows-Security Module/Main files/C#/Others/IniFileConverter.cs
