+++ Harden-Windows-Security Module/Main files/C#/Others/Initializer.cs
