+++ Harden-Windows-Security Module/Main files/C#/Others/JsonToHashtable.cs
