+++ Harden-Windows-Security Module/Main files/C#/Others/LocalGroupMember.cs
