+++ Harden-Windows-Security Module/Main files/C#/Others/Logger.cs
