+++ Harden-Windows-Security Module/Main files/C#/Others/MDMClassProcessor.cs
