+++ Harden-Windows-Security Module/Main files/C#/Others/Miscellaneous.cs
