+++ Harden-Windows-Security Module/Main files/C#/Others/MitigationPolicyProcessor.cs
