+++ Harden-Windows-Security Module/Main files/C#/Others/NewToastNotification.cs
