+++ Harden-Windows-Security Module/Main files/C#/Others/PowerShellExecutor.cs
