+++ Harden-Windows-Security Module/Main files/C#/Others/ProcessMitigationsApplication.cs
