+++ Harden-Windows-Security Module/Main files/C#/Others/ProcessMitigationsParser.cs
