+++ Harden-Windows-Security Module/Main files/C#/Others/PropertyHelper.cs
