+++ Harden-Windows-Security Module/Main files/C#/Others/RegistryEditor.cs
