+++ Harden-Windows-Security Module/Main files/C#/Others/RunCommand.cs
