+++ Harden-Windows-Security Module/Main files/C#/Others/SecureStringComparer.cs
