+++ Harden-Windows-Security Module/Main files/C#/Others/SecuriryPolicyProcessor.cs
