+++ Harden-Windows-Security Module/Main files/C#/Others/SneakAndPeek.cs
