+++ Harden-Windows-Security Module/Main files/C#/Others/SystemInfoNativeMethods.cs
