+++ Harden-Windows-Security Module/Main files/C#/Others/UserPrivCheck.cs
