+++ Harden-Windows-Security Module/Main files/C#/Others/VerboseLogger.cs
