+++ Harden-Windows-Security Module/Main files/C#/Others/WindowsFeatureChecker.cs
